KAPT: Move compiler plugin and AP sources from resources
to the integration test itself.
 #KT-80882
