--- libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/resources/testProject/kapt2/compilerPluginFunctionKind/build.gradle
@@ -1,1 +1,0 @@
-// Root build script intentionally empty.
