--- libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/resources/testProject/kapt2/compilerPluginFunctionKind/example/build.gradle
@@ -1,1 +1,0 @@
-// Configured from KaptCompilerPluginsIT via build script injection.
