--- libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/resources/testProject/kapt2/compilerPluginFunctionKind/example/src/main/java/Reproducer.kt
@@ -1,27 +1,0 @@
-package repro
-
-import repro.processor.Trigger
-
-@Target(AnnotationTarget.TYPE)
-annotation class PluginFunction
-
-@Trigger
-interface TestInterface {
-    fun test(block: @PluginFunction () -> Unit)
-
-    fun testList(blocks: List<@PluginFunction () -> Unit>)
-
-    fun testReturn(): @PluginFunction () -> Unit
-
-    fun testListReturn(): List<@PluginFunction () -> Unit>
-}
-
-class TestClass {
-    val directProperty: @PluginFunction () -> Unit = {}
-
-    val listProperty: List<@PluginFunction () -> Unit> = emptyList()
-
-    var mutableDirectProperty: @PluginFunction () -> Unit = {}
-
-    var mutableListProperty: List<@PluginFunction () -> Unit> = emptyList()
-}
