--- libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/resources/testProject/kapt2/compilerPluginFunctionKind/settings.gradle
@@ -1,1 +1,0 @@
-include ':annotation-processor', ':example', ':plugin'
